fixes #18

+++ main.js
@@ -614,9 +614,7 @@ class Robonect extends utils.Adapter {
                 rule,
                 data
             );
-            if (objects[item].common.type === 'number') val=Number.parseInt(val);
-            // if (objects[item].common.type === 'boolean' && item === 'status.stopped') val= val===1;
-            // this.log.debug(`Trying to set ${item} to value: [${val}]`);
+            if (objects[item].common.type === 'number') val=Number.parseFloat(val);
             this.setState(item, { val: val, ack: true });
         }
     }
